@@ -1,53 +1,58 @@
 {
-  "name": "starter",
-  "version": "0.0.0",
+  "name": "essentials",
+  "version": "3.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve -o",
+    "start": "ng serve",
+    "proxy": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "tsc": "tsc"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "11.0.2",
-    "@angular/common": "11.0.2",
-    "@angular/compiler": "11.0.2",
-    "@angular/core": "11.0.2",
-    "@angular/forms": "11.0.2",
-    "@angular/platform-browser": "11.0.2",
-    "@angular/platform-browser-dynamic": "11.0.2",
-    "@angular/router": "11.0.2",
-    "angular-oauth2-oidc": "^10.0.3",
-    "bootstrap": "^4.0.0-beta.2",
-    "core-js": "^2.5.1",
-    "rxjs": "^6.5.4",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "@angular/animations": "17.3.3",
+    "@angular/common": "17.3.3",
+    "@angular/compiler": "17.3.3",
+    "@angular/core": "17.3.3",
+    "@angular/forms": "17.3.3",
+    "@angular/platform-browser": "17.3.3",
+    "@angular/platform-browser-dynamic": "17.3.3",
+    "@angular/router": "17.3.3",
+    "bootstrap": "^4.6.2",
+    "rxjs": "^7.8.1",
+    "tslib": "~2.3.1",
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular/cli": "11.0.2",
-    "@angular/compiler-cli": "11.0.2",
-    "@angular/language-service": "11.0.2",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~3.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "4.0.5",
-    "@angular-devkit/build-angular": "~0.1100.2"
+    "@angular-builders/jest": "17.0.3",
+    "@angular-devkit/build-angular": "17.3.3",
+    "@angular-eslint/builder": "17.3.0",
+    "@angular-eslint/eslint-plugin": "17.3.0",
+    "@angular-eslint/eslint-plugin-template": "17.3.0",
+    "@angular-eslint/schematics": "17.3.0",
+    "@angular-eslint/template-parser": "17.3.0",
+    "@angular/cli": "~17.3.3",
+    "@angular/compiler-cli": "17.3.3",
+    "@angular/language-service": "17.3.3",
+    "@briebug/jest-schematic": "^6.0.0",
+    "@cypress/schematic": "^2.5.1",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.12.4",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "cypress": "^13.7.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "^3.2.5",
+    "typescript": "~5.2.2",
+    "yarn-upgrade-all": "^0.7.2"
   }
 }
