@@ -8,12 +8,6 @@ if (environment.production) {
   enableProdMode();
 }
 
-
-
-
-
-
-
-
-
-platformBrowserDynamic().bootstrapModule(AppModule);
+platformBrowserDynamic()
+  .bootstrapModule(AppModule)
+  .catch((err) => console.error(err));
